Move SanitizerSet to rustc_target
